Relocate iwshader sources and drop r_lightmap_strength
