+++ Quake/Makefile.w32
@@ -268,7 +268,7 @@ OBJS = strlcat.o \
         wad.o \
         cmd.o \
         common.o \
-        common/com_iwtext.o \
+        com_iwtext.o \
         steam.o \
 	json.o \
 	miniz.o \
