+++ Quake/gl_screen.c
@@ -24,7 +24,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 // screen.c -- master for refresh, status bar, console, chat, notify, etc
 
 #include "quakedef.h"
-#include "renderer/r_iwshader.h"
+#include "r_iwshader.h"
 #include "steam.h"
 #include <time.h>
