+++ Quake/r_world.c
@@ -24,7 +24,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
 #include "quakedef.h"
 #include "gl_texmgr.h"
-#include "renderer/r_iwshader.h"
+#include "r_iwshader.h"
 #include "image.h"
 
 #include <math.h>
