+++ Windows/VisualStudio/ironwail.vcxproj
@@ -234,7 +234,7 @@ copy "$(SolutionDir)..\zlib\$(PlatformShortName)\*.dll" "$(TargetDir)"</Command>
     <ClCompile Include="..\..\Quake\cl_tent.c" />
     <ClCompile Include="..\..\Quake\cmd.c" />
     <ClCompile Include="..\..\Quake\common.c" />
-    <ClCompile Include="..\..\Quake\common\com_iwtext.c" />
+    <ClCompile Include="..\..\Quake\com_iwtext.c" />
     <ClCompile Include="..\..\Quake\console.c" />
     <ClCompile Include="..\..\Quake\crc.c" />
     <ClCompile Include="..\..\Quake\cvar.c" />
@@ -300,7 +300,7 @@ copy "$(SolutionDir)..\zlib\$(PlatformShortName)\*.dll" "$(TargetDir)"</Command>
     <ClCompile Include="..\..\Quake\r_part.c" />
     <ClCompile Include="..\..\Quake\r_sprite.c" />
     <ClCompile Include="..\..\Quake\r_world.c" />
-    <ClCompile Include="..\..\Quake\renderer\r_iwshader.c" />
+    <ClCompile Include="..\..\Quake\r_iwshader.c" />
     <ClCompile Include="..\..\Quake\sbar.c" />
     <ClCompile Include="..\..\Quake\snd_codec.c" />
     <ClCompile Include="..\..\Quake\snd_dma.c" />
@@ -430,7 +430,7 @@ copy "$(SolutionDir)..\zlib\$(PlatformShortName)\*.dll" "$(TargetDir)"</Command>
     <ClInclude Include="..\..\Quake\client.h" />
     <ClInclude Include="..\..\Quake\cmd.h" />
     <ClInclude Include="..\..\Quake\common.h" />
-    <ClInclude Include="..\..\Quake\common\com_iwtext.h" />
+    <ClInclude Include="..\..\Quake\com_iwtext.h" />
     <ClInclude Include="..\..\Quake\console.h" />
     <ClInclude Include="..\..\Quake\crc.h" />
     <ClInclude Include="..\..\Quake\cvar.h" />
@@ -466,7 +466,7 @@ copy "$(SolutionDir)..\zlib\$(PlatformShortName)\*.dll" "$(TargetDir)"</Command>
     <ClInclude Include="..\..\Quake\q_sound.h" />
     <ClInclude Include="..\..\Quake\q_stdinc.h" />
     <ClInclude Include="..\..\Quake\render.h" />
-    <ClInclude Include="..\..\Quake\renderer\r_iwshader.h" />
+    <ClInclude Include="..\..\Quake\r_iwshader.h" />
     <ClInclude Include="..\..\Quake\resource.h" />
     <ClInclude Include="..\..\Quake\sbar.h" />
     <ClInclude Include="..\..\Quake\screen.h" />
