+++ Windows/VisualStudio/ironwail.vcxproj.filters
@@ -45,7 +45,7 @@
     <ClCompile Include="..\..\Quake\common.c">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\Quake\common\com_iwtext.c">
+    <ClCompile Include="..\..\Quake\com_iwtext.c">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="..\..\Quake\console.c">
@@ -174,7 +174,7 @@
     <ClCompile Include="..\..\Quake\r_world.c">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\Quake\renderer\r_iwshader.c">
+    <ClCompile Include="..\..\Quake\r_iwshader.c">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="..\..\Quake\sbar.c">
@@ -350,7 +350,7 @@
     <ClInclude Include="..\..\Quake\common.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\Quake\common\com_iwtext.h">
+    <ClInclude Include="..\..\Quake\com_iwtext.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="..\..\Quake\console.h">
@@ -449,7 +449,7 @@
     <ClInclude Include="..\..\Quake\render.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\Quake\renderer\r_iwshader.h">
+    <ClInclude Include="..\..\Quake\r_iwshader.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="..\..\Quake\resource.h">
